backend/requirements.txt
# MEXAR Phase 1 - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Groq API
groq==0.4.2
httpx==0.27.0  # Pin to compatible version for groq SDK

# Knowledge Graph
networkx==3.2.1

# Data Processing
pandas==2.1.4
PyPDF2==3.0.1
python-docx==1.1.0

# File Upload
python-multipart==0.0.6

# Video Processing
opencv-python==4.9.0.80

# Environment
python-dotenv==1.0.0

# JSON handling
orjson==3.9.10

# Async support
aiofiles==23.2.1

# Database (Supabase/PostgreSQL)
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
email-validator==2.1.0

# Supabase Client
supabase==2.24.0

# Vector Support
fastembed>=0.7.0  # Updated from 0.2.0 (was yanked)
pgvector==0.2.4

# RAG Components (NEW)
sentence-transformers>=2.2.0  # Cross-encoder reranking
numpy>=1.24.0  # Vector operations
